requirements.txt
chromadb==1.5.5
langchain==1.2.13
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.55.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_utils==0.1.0